+++ integrations/github/github/core/options.py
@@ -25,6 +25,7 @@ class ListRepositoryOptions(SingleOrganizationOptions):
     organization_type: Required[str]
     search_params: NotRequired[Optional[RepoSearchParams]]
     included_relations: NotRequired[Optional[dict[str, dict[str, Any]]]]
+    updated_since: NotRequired[Optional[datetime]]
 
 
 class RepositoryIdentifier(SingleOrganizationOptions):
@@ -79,6 +80,7 @@ class ListIssueOptions(RepositoryIdentifier):
 
     state: Required[str]
     labels: NotRequired[Optional[str]]
+    since: NotRequired[Optional[datetime]]
 
 
 class BaseUserOptions(SingleOrganizationOptions):
@@ -130,6 +132,8 @@ class SingleReleaseOptions(RepositoryIdentifier):
 
 class ListReleaseOptions(RepositoryIdentifier):
     """Options for listing releases."""
+
+    created_since: NotRequired[Optional[datetime]]
 
 
 class SingleTagOptions(RepositoryIdentifier):
@@ -183,6 +187,7 @@ class ListDeploymentsOptions(RepositoryIdentifier):
     task: NotRequired[Optional[str]]
     environment: NotRequired[Optional[str]]
     enrich_with_first_commit: NotRequired[bool]
+    created_since: NotRequired[Optional[datetime]]
 
 
 class SingleDeploymentStatusOptions(RepositoryIdentifier):
@@ -210,6 +215,7 @@ class ListDependabotAlertOptions(RepositoryIdentifier):
     state: Required[list[str]]
     severity: NotRequired[Optional[str]]
     ecosystem: NotRequired[Optional[str]]
+    updated_since: NotRequired[Optional[datetime]]
 
 
 class SingleCodeScanningAlertOptions(RepositoryIdentifier):
@@ -223,6 +229,7 @@ class ListCodeScanningAlertOptions(RepositoryIdentifier):
 
     state: Required[str]
     severity: NotRequired[Optional[str]]
+    updated_since: NotRequired[Optional[datetime]]
 
 
 class FileContentOptions(RepositoryIdentifier):